+++ src/widgets/TopSources.php
@@ -9,7 +9,7 @@
 namespace shornuk\plausible\widgets;
 
 use shornuk\plausible\Plausible;
-use shornuk\plausible\services\PlausibleService;
+use shornuk\plausible\helpers\StringHelper;
 use shornuk\plausible\assetbundles\plausible\PlausibleAsset;
 
 use Craft;
@@ -28,8 +28,8 @@ class TopSources extends Widget
     // Public Properties
     // =========================================================================
 
-    public $limit = 4;
-    public $timePeriod = '30d';
+    public int $limit = 4;
+    public string $timePeriod = '30d';
 
     // Static Methods
     // =========================================================================
@@ -67,17 +67,9 @@ class TopSources extends Widget
 
     public function getTitle(): ?string
     {
-        if (!isset($title)) {
-            $title = Craft::t('plausible', 'Top Sources');
-        }
-        $timePeriod = $this->timePeriod;
-
-        if ($timePeriod) {
-            $title = Craft::t('app', 'Top Sources - {timePeriod}', [
-                'timePeriod' => Craft::t('plausible', Plausible::$plugin->plausible->timeLabelize($timePeriod)),
-            ]);
-        }
-        return $title;
+        return Craft::t('plausible', 'Top Sources - {timePeriod}', [
+            'timePeriod' => Craft::t('plausible', StringHelper::timeLabelize($this->timePeriod)),
+        ]);
     }
 
     /**
@@ -100,11 +92,29 @@ class TopSources extends Widget
     {
         Craft::$app->getView()->registerAssetBundle(PlausibleAsset::class);
 
-        $cacheKey = 'plausible:topSources'.$this->timePeriod.$this->limit;
+        $cacheKey = 'plausibleV5:topSources'.$this->timePeriod.$this->limit;
         $results = Craft::$app->getCache()->get($cacheKey);
         if (!$results)
         {
-            $results = Plausible::$plugin->plausible->getTopSources($this->limit, $this->timePeriod);
+
+            $results = Plausible::$plugin->plausible->query([
+                'metrics' => ['visitors'],
+                'date_range' => $this->timePeriod,
+                'pagination' => [
+                    'limit' => $this->limit,
+                    'offset' => 0,
+                ],
+                "filters" => [
+                    [
+                        "is_not",
+                        "visit:source",
+                        [""]
+                    ]
+                ],
+                "dimensions" => [
+                    "visit:source"
+                ]
+            ]);
             Craft::$app->getCache()->set($cacheKey, $results, 300);
         }